Start extracting device auth for an rpc experiment
